% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3_check.R
\name{check.ztu}
\alias{check.ztu}
\title{check floodmap 6.3 data. Don't suggest, will be removed in future.}
\usage{
check.ztu(indir = NULL, fnames)
}
\arguments{
\item{fnames}{if inputdir is null, fnames will be used.}

\item{inputdir}{6.3/ztu path}
}
\description{
before use this function, should be check weather there have value field. Otherwise, will
generate fatal error
}
\examples{
TMP <- ztu.check(indir)
}

