% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3_check.R
\name{check.ztu_formatDBF}
\alias{check.ztu_formatDBF}
\title{format 6.3 ztu dbf field type, precision, scale}
\usage{
check.ztu_formatDBF(indir)
}
\arguments{
\item{indir}{6.3/ztu path}
}
\description{
only used for ztu shapefile data. Out of date, don't suggest
}
\examples{
TMP <- check.ztu_formatDBF(indir)
}

