% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_smallGrid.R
\name{get_maskshp}
\alias{get_maskshp}
\title{get_maskshp}
\usage{
get_maskshp(ztu, fn = "maskshp", outdir = "FloodScheme/", onlyYMSS = T)
}
\arguments{
\item{onlyYMSS}{use ymss.shp or ddsj, ymls, ymss to clip}
}
\value{
maskshp.shp at outdir directory
}
\description{
based ztu directory shapefiles clip grid.shp in 6.2
}
