% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_smallGrid.R
\name{get_schemeshp}
\alias{get_schemeshp}
\title{schemeshp}
\usage{
get_schemeshp(ztu, outdir = "FloodScheme/")
}
\value{
maskshp.shp at outdir directory
}
\description{
according to 6.3 ztu ddsjn.shp, ymlsn.shp, ymssn.shp to write flood scheme shp files.
SHOULD BE ATTENTION ZTU directory can't contain shpfiles except ddsjn.shp, ymlsn.shp, ymssn.shp, and
the three shp must in ztu directory. Finally generate scheme shapefiles as MIKE II model
export, to floodmapping system.
}
