% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_modify.R
\name{prj.check}
\alias{prj.check}
\title{prj.check}
\usage{
prj.check(indir)
}
\arguments{
\item{indir}{the six part directory}
}
\description{
check projection is or not CGCS_2000
}
\examples{
prj.check(indir)
}

