% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbf.R
\name{readDBF}
\alias{readDBF}
\title{read dbf file}
\usage{
read.dbf(file, as.is = FALSE)
}
\arguments{
\item{file}{name of input file}

\item{as.is}{should character vectors not be converted to factors?}
}
\description{
The function reads a DBF file into a data frame, converting character
fields to factors, and trying to respect NULL fields.
The DBF format is documented but not much adhered to.  There is is no
guarantee this will read all DBF files.
}
\examples{
 x <- read.dbf(system.file("files/sids.dbf", package="foreign")[1])
 str(x)
 summary(x)
}
\references{
url{http://shapelib.maptools.org/}.
 The Borland file specification \emph{via} \url{http://www.wotsit.org},
 currently at \url{http://www.wotsit.org/list.asp?fc=6}.
}
\seealso{
\code{\link{write.dbf}}
}

