% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlterField.R
\name{readDBF_fields}
\alias{readDBF_fields}
\title{readDBF_fields}
\usage{
readDBF_fields(dbf.name)
}
\arguments{
\item{dbf.name}{dbf file name}
}
\details{
read dbf table fields Name, TYPE, precision and scale. The result is same as ArcGIS
}
\examples{
fnames <- dir(system.file("data", package = "floodmap"), pattern = "*.dbf$", full.names = T)
for (i in seq_along(fnames)){
   cat(sprintf("[\%02d]----\%s\\n", i, fnames[i]))
   print(readDBF_fields(fnames[i]))
}
}
