% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfWrite_xlsx.R
\name{readDBF_List2xlsx}
\alias{readDBF_List2xlsx}
\title{readDBF_List2xlsx}
\usage{
readDBF_List2xlsx(indir, fname, save = TRUE, as.is = FALSE)
}
\arguments{
\item{indir}{directory which shapefiles dbf file store}

\item{fname}{file name which dbf data saved}

\item{save}{if save is true, save into file, else import to R variable environment}
}
\description{
read shapefile dbf table and write into excel *.xlsx
}
\examples{
inputdir <- system.file("data", package = "floodmap")
reaDBF_List2xlsx(inputdir, "test01.xlsx")
}

