% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlterField.R
\name{readDBFs_fields}
\alias{readDBFs_fields}
\title{readDBFs_fields}
\usage{
readDBFs_fields(indir, fnames = NULL, checkField = FALSE, show = TRUE)
}
\arguments{
\item{indir}{a directory under which have dbf files}
}
\details{
read dbf table fields Name, TYPE, precision and scale. The result is same as ArcGIS
if fnames input, indir will be ignore
}
\examples{
inputdir <- system.file("data", package = "floodmap")
fields <- readDBFs_fields(inputdir)
}

