% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbf.R
\name{writeDBF}
\alias{writeDBF}
\title{writeDBF}
\usage{
writeDBF(dataframe, file, factor2char = TRUE, max_nchar = 254,
  precisionIn = NULL, scaleIn = NULL, copyfile = F)
}
\arguments{
\item{precisionIn}{filed length, should be caution that: number of digits including minus sign and decimal sign
scaleIn: digits after the decimal sign}

\item{scaleIn}{numbers after digital point.}

\item{cofyfile}{if TRUE, original dbf file would be rename to "*.dbfold" to present dbf data missing}
}
\description{
modified write.dbf from foreign package, in order to control field precision
 and scale properties
}

